Time: 3 ms (50.27%), Space: 7.5 MB (55.71%) - LeetHub

+++ 0791-custom-sort-string/0791-custom-sort-string.cpp
@@ -1,8 +1,8 @@
 class Solution {
 public:
     string customSortString(string o, string s) {
-        vector<int> v(26, 27);
-        int cnt = 0;
+        vector<short> v(26, 27);
+        short cnt = 0;
         for(char c : o) {
             v[c - 'a'] = cnt++;
         }
